backend/requirements.txt
# For FastAPI test client
httpx
# Testing
pytest
# Pre-commit and code quality
pre-commit
black
flake8
flake8-bugbear
# Secret scanning
detect-secrets
# Backend Python requirements

# Core
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv

# Auth & Security
passlib[bcrypt]
bcrypt==4.1.2
python-jose[cryptography]
python-multipart

# Pydantic (already required by FastAPI)

# Pagination, Filtering, Sorting
fastapi-pagination

# File Uploads
aiofiles

# Email Notifications
aiosmtplib

# Required for Pydantic email validation
email-validator

# Admin/Management
alembic

# Logging
loguru

# Testing
requests
